Fix: Assignment Lecture 07 and 08
- Fix Assignment 02 Lecture 07
- Fix Assignment 02 Lecture 08
- Fix Assignment 03 Lecture 08

+++ Week 04/Lecture 08/Assignment 03/assignment3/src/main/java/example/lecture8/assignment3/repository/EmployeeRepository2.java
@@ -1,42 +1,35 @@
 package example.lecture8.assignment3.repository;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.dao.DataIntegrityViolationException;
 
 import example.lecture8.assignment3.model.Employee;
 
 @Repository
 public class EmployeeRepository2 {
     private final JdbcTemplate jdbcTemplate;
+    private final EmployeeRowMapper employeeRowMapper;
 
-    public EmployeeRepository2(@Qualifier("jdbcTemplate2") JdbcTemplate jdbcTemplate) {
+    public EmployeeRepository2(@Qualifier("jdbcTemplate2") JdbcTemplate jdbcTemplate, EmployeeRowMapper employeeRowMapper) {
         this.jdbcTemplate = jdbcTemplate;
-    }
-
-    private static final class EmployeeRowMapper implements RowMapper<Employee> {
-        @Override
-        public Employee mapRow(ResultSet rs, int rowNum) throws SQLException {
-            Employee emp = new Employee();
-            emp.setId(rs.getInt("id"));
-            emp.setName(rs.getString("name"));
-            emp.setDepartment(rs.getString("department"));
-            emp.setAge(rs.getInt("age"));
-            return emp;
-        }
+        this.employeeRowMapper = employeeRowMapper;
     }
 
     public List<Employee> findAll() {
-        return jdbcTemplate.query("SELECT * FROM employee", new EmployeeRowMapper());
+        return jdbcTemplate.query("SELECT * FROM employee", employeeRowMapper);
     }
 
     public Employee findEmployeeById(int id) {
-        return jdbcTemplate.queryForObject("SELECT * FROM employee WHERE id = ?", new EmployeeRowMapper(), id);
+        try {
+            return jdbcTemplate.queryForObject("SELECT * FROM employee WHERE id = ?", employeeRowMapper, id);
+        } catch(EmptyResultDataAccessException ex) {
+            return null;
+        }
     }
 
     public int createEmployee(Employee emp) {
@@ -44,10 +37,22 @@ public class EmployeeRepository2 {
     }
 
     public int updateEmployee(Employee emp, int id) {
-        return jdbcTemplate.update("UPDATE employee SET name = ?, department = ?, age = ? WHERE id = ?", emp.getName(), emp.getDepartment(), emp.getAge(), id);
+        try {
+            if(findEmployeeById(id) == null) {
+                return -1;
+            }
+            
+            return jdbcTemplate.update("UPDATE employee SET name = ?, department = ?, age = ? WHERE id = ?", emp.getName(), emp.getDepartment(), emp.getAge(), id);
+        } catch(DataIntegrityViolationException ve) {
+            return -2;
+        }
     }
 
     public int deleteEmployee(int id) {
+        if(findEmployeeById(id) == null) {
+            return -1;
+        }
+        
         return jdbcTemplate.update("DELETE FROM employee WHERE id = ?", id);
     }
 }
